Refactor code to pass strict (based)pyright While this code has worked for a while to pass the strict settings of MyPy, unfortunately MyPy has lagged quite a bit behind in recent times. Since I code in VSCodium, I can't get access to Pylance as Microsoft deliberately codes their extensions to only work on VSCode.
This is where I had to turn to "basedpyright", a fork of Pyright that is meant to reincorporate a lot of the features of Pylance. That being said, basedpyright is really strict about how it wants you to write your code. Some of them are so verbose/pedantic that I will ignore them.

This commit fixes missing type hints that I didn't see earlier; removes `Any` type hints by declaring additional typedefs (e.g. JSON_Value), specifies the proper types, and in the worst cases, specify to the "object" type and `cast` to the proper type for those really nasty warnings; replaces deprecated type hints with modern ones; adds decorators for method overrides; uses a new method of optional defaults in constructors

In addition, a couple of bugs were patched that I did not see previously, regarding values not being converted to JSON dictionaries when serializing.